requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
pydantic[email]==2.8.2
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0
redis==5.0.7
rq==1.16.2
minio==7.2.7
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
anyio==4.4.0
fakeredis==2.23.2
email-validator==2.2.0
idna==3.7
